pull buildSelector func into own file
